fix related to long Id

+++ ResumableFunctions.Handler/UiService/IUiService.cs
@@ -6,18 +6,18 @@ namespace ResumableFunctions.Handler.UiService
     public interface IUiService
     {
         Task<MainStatistics> GetMainStatistics();
-        Task<ServiceData> GetServiceInfo(int serviceId);
-        Task<List<LogRecord>> GetServiceLogs(int serviceId);
+        Task<ServiceData> GetServiceInfo(long serviceId);
+        Task<List<LogRecord>> GetServiceLogs(long serviceId);
         Task<List<LogRecord>> GetLogs(int page = 0);
         Task<List<ServiceInfo>> GetServicesList();
-        Task<ServiceStatistics> GetServiceStatistics(int serviceId);
-        Task<List<FunctionInfo>> GetFunctionsInfo(int? serviceId);
-        Task<List<MethodGroupInfo>> GetMethodsInfo(int? serviceId);
-        Task<List<PushedCallInfo>> GetPushedCalls(int page);
-        Task<List<FunctionInstanceInfo>> GetFunctionInstances(int functionId);
-        Task<PushedCallDetails> GetPushedCallDetails(int pushedCallId);
-        Task<FunctionInstanceDetails> GetInstanceDetails(int instanceId);
-        Task<List<MethodInGroupInfo>> GetMethodsInGroup(int groupId);
-        Task<List<MethodWaitDetails>> GetWaitsForGroup(int groupId);
+        Task<ServiceStatistics> GetServiceStatistics(long serviceId);
+        Task<List<FunctionInfo>> GetFunctionsInfo(long? serviceId);
+        Task<List<MethodGroupInfo>> GetMethodsInfo(long? serviceId);
+        Task<List<PushedCallInfo>> GetPushedCalls(long page);
+        Task<List<FunctionInstanceInfo>> GetFunctionInstances(long functionId);
+        Task<PushedCallDetails> GetPushedCallDetails(long pushedCallId);
+        Task<FunctionInstanceDetails> GetInstanceDetails(long instanceId);
+        Task<List<MethodInGroupInfo>> GetMethodsInGroup(long groupId);
+        Task<List<MethodWaitDetails>> GetWaitsForGroup(long groupId);
     }
 }
